Moves validate_stream() to test_helper.c
This enables running validation in several files.
Validation of FH-TG has been added.
